requirements.txt
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
PyJWT==2.9.0
mysql-connector-python==9.0.0
